% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{B_all}
\alias{B_all}
\title{Local beta coefficients from source sites}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 25 rows and 16 columns.
}
\usage{
B_all
}
\description{
A 25*16 numeric matrix containing the local estimated beta coefficients from all 16 source sites.
Each would become a target source under the federated learning framework. Rows stand for different
variables while columns mean different sites.
}
\keyword{datasets}
