% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{B_source}
\alias{B_source}
\title{Local beta coefficients from source sites}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 25 rows and 15 columns.
}
\usage{
B_source
}
\description{
A 25*15 numeric matrix containing the local estimated beta coefficients from 15 source sites.
Rows stand for different variables while columns mean different source sites.
}
\keyword{datasets}
