% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta.valid}
\alias{delta.valid}
\title{Status indicator for the validation dataset.}
\format{
An object of class \code{numeric} of length 2000.
}
\usage{
delta.valid
}
\description{
A 0-1 vector indicating the status of 2000 patients in the validation dataset. 1=dead, 0=censored.
}
\keyword{datasets}
