% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_all}
\alias{Sigma_all}
\title{Data covariance matrix from each site.}
\format{
An object of class \code{list} of length 16.
}
\usage{
Sigma_all
}
\description{
A list containing 16 elements, each of which being a 25*25 covariance
matrix calculated from one site.
}
\keyword{datasets}
