% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_target}
\alias{Sigma_target}
\title{Data covariance matrix from the target site.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 25 rows and 25 columns.
}
\usage{
Sigma_target
}
\description{
A 25*25 covariance matrix calculated from the target site.
}
\keyword{datasets}
