% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_maximin.R
\name{survmaximin}
\alias{survmaximin}
\title{Generate transfer-learning survmaximin coefficients}
\usage{
survmaximin(B_source, Sigma_target, delta = 0)
}
\arguments{
\item{B_source}{A p (number of variables) * L (number of source sites) numeric matrix.}

\item{Sigma_target}{Data covariance matrix for the target site.}

\item{delta}{Default 0; the ridge penalty parameter.}
}
\value{
A list with the following components:

\tabular{ll}{
 \code{beta.est} \tab Estimated beta coefficients for the target site. \cr
 \code{weight} \tab A vector containing trained weights for each source site. \cr
}
}
\description{
Generate transfer-learning survmaximin coefficients
}
\examples{
data(B_source); data(Sigma_target)
output <- survmaximin(B_source, Sigma_target, delta=0.5)
beta.maximin <- output$beta.est
}
