% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_maximin.R
\name{survmaximin_fed}
\alias{survmaximin_fed}
\title{Generate federate-learning coefficients}
\usage{
survmaximin_fed(B_all, Sigma_all, delta = 0)
}
\arguments{
\item{B_all}{A p (number of variables) * L (number of sites) numeric matrix.}

\item{Sigma_all}{A list containing the data covariance matrix for each site.}

\item{delta}{Default 0; the ridge penalty parameter.}
}
\value{
A list with each element as below, being the estimated results for each site:

\tabular{ll}{
 \code{beta.est} \tab Estimated beta coefficients for the target site. \cr
 \code{weight} \tab A vector containing trained weights for each source site. \cr
}
}
\description{
Generate federate-learning coefficients
}
\examples{
data(B_all); data(Sigma_all)
output <- survmaximin_fed(B_all, Sigma_all, delta=0.5)
beta.maximin <- output$beta.est
}
