% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x.valid}
\alias{x.valid}
\title{Time to event for the validation dataset.}
\format{
An object of class \code{numeric} of length 2000.
}
\usage{
x.valid
}
\description{
A numeric vector denoting the followup time for 2000 patients in the validation dataset.
}
\keyword{datasets}
