% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{z.valid}
\alias{z.valid}
\title{Baseline covariates in the validation dataset.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2000 rows and 25 columns.
}
\usage{
z.valid
}
\description{
A numeric matrix including 25 baseline covariates for the validation dataset.
}
\keyword{datasets}
